Add files via upload
This is the basic barebones platformer game. It includes jumping and movement, as well as gravity, but it does have issues with the camera following the player. The player for some reason gets stuck in the bottom-most platform(ground). Also 2 of the other platforms allow for the player to jump off of them however when the player should stay on top of them the ball instead falls back to the ground. Also need to fix ball teleporting to top of platform if touching while jumping.